refactor(chain): RPC-failover read-intent + stickiness ergonomics (#1548) (#1560)
* refactor(chain): RPC-failover read-intent + stickiness ergonomics (#1548)

Behavior-preserving follow-up to #1544. No production behavior change; all
packages/chain unit tests (600 pass / 1 skip), mock-adapter parity, and tsc
pass unchanged — the only test diff is relocating the Mechanism-B AC suite.

- hub-rotation-poller: fold the 3 background tip probes behind a `readTip`
  wrapper so no call site spells the transport-internal `skipPreferred` flag.
- endpoint-stickiness: add `attempts(canonical, intent)` bundling the per-loop
  order + recordSuccess + recordFailure triple; rewire all 4 failover loops so
  (canonical, intent) can't drift across the three calls. order() still runs
  exactly once per loop entry.
- evm-adapter-conviction: collapse the PCA requestPublishingConvictionRpc
  4-branch method/params ladder into a pure classifyPcaRead classifier + one
  dispatch (identical per-method routing).
- test: split the ~580-line endpoint-stickiness AC suite (AC-1..AC-29) out of
  rpc-failover-client.unit.test.ts into a dedicated acceptance file.


* refactor(chain): seal stickiness attempt API + exhaustive PCA dispatch (#1548 review)

Addresses the two otReviewAgent comments on #1560 (both behavior-preserving):

- endpoint-stickiness: mark `order`/`recordSuccess`/`recordFailure` `private` so
  `attempts()` is the SOLE public driving API — the mismatched-(canonical, intent)
  hand-threaded path the bundle prevents is now unavailable, not just discouraged.
  Only caller is `attempts()` itself; the client already routes through `plan.*`.
  TS `private` is erased at runtime, so behavior + the prototype are unchanged.
- evm-adapter-conviction: make the PCA `classifyPcaRead` dispatch exhaustive —
  four explicit cases + a `const _exhaustive: never = strategy` tail, so a future
  strategy that forgets a transport is a compile error, not a silent readProvider.
- test: drop the now-dead `getRpcFailoverStats` import left in
  rpc-failover-client.unit.test.ts after the AC block moved out (item 4).

Fast lane 600 pass / 1 skip, parity 15, tsc clean — all unchanged.


* refactor(chain): bind the stickiness attempt (endpoint + outcome) not just the plan (#1548 review)

Addresses otReviewAgent's follow-up: push the abstraction one step further so the
endpoint/index coupling is structural, not spread across every failover loop.

`EndpointStickiness.attempts(canonical, intent)` now returns an array of BOUND
`StickyAttempt` entries — `{ endpoint, recordSuccess(), recordFailure() }` — with
the endpoint, its position (triedFirst = index 0), and the (canonical, intent)
pair all captured at creation. The four failover loops iterate the entries and
call the NULLARY `attempts[i].recordSuccess()` / `attempts[i].recordFailure()` on
the one they ran, so recording against the wrong endpoint, a stale index, or a
mismatched canonical/intent is now unrepresentable.

Behavior-identical: `order()` still runs exactly once (attempts maps over its
result); each entry expands to the prior `recordSuccess(endpoints[i], canonical,
intent, i===0)` / `recordFailure(endpoints[i], intent)`; closures capture
`canonical` (not the reordered list) for the primary check; all record fire-points
unchanged. The residual loop `i` is only failover telemetry (next-hop rpcUrl) +
the last-attempt check — deliberately kept out of the pure state machine.

Fast lane 600 pass / 1 skip, parity 15, tsc clean — all unchanged.


* refactor(chain): drive failover loops directly from bound attempts (#1548 review)

Addresses otReviewAgent: the bound-attempt abstraction was immediately flattened
back into an index-coupled parallel array (`endpoints = attempts.map(a=>a.endpoint)`
used as `endpoints[i]` beside `attempts[i]`), which re-created the coupling the
bind was meant to remove.

All four failover loops now iterate `attempts` as the authoritative item — the
parallel `endpoints` array is gone:

  const attempt = attempts[i];
  const endpoint = attempt.endpoint;
  ... attempt.recordSuccess() / attempt.recordFailure() ...
  noteRpcFailover(label, endpoint.rpcUrl, err, attempts[i + 1].endpoint.rpcUrl);

The endpoint and its recorder stay structurally coupled through the whole pass;
there is no second array to drift against. In runAcrossProviders the pre-existing
local read-promise `attempt` was renamed to `pending` to free the name. The
residual index is only the last-attempt check + the next-hop failover-log URL,
deliberately kept out of the pure state machine.

Behavior-identical (endpoints[i] === attempts[i].endpoint, endpoints.length ===
attempts.length, every record fire-point unchanged): fast lane 600 pass / 1 skip,
parity 15, tsc clean, plus an independent two-lens equivalence review.


* refactor(chain): method-exhaustive PCA classifier + shared failover test harness (#1548 review)

Addresses two otReviewAgent round-4 comments (both behavior/assertion-preserving):

- evm-adapter-conviction: `classifyPcaRead` is now an exhaustive `switch (method)`
  over `PcaRpcMethod` with a `const _exhaustive: never = method` tail, so a newly
  added PCA method must make a conscious routing decision at compile time (TS2322)
  instead of silently falling through to `sticky`. Routing for all six existing
  methods is byte-identical; the now-redundant `PCA_NULLABLE_LOOKUP_METHODS` set is
  removed (the switch encodes nullable routing explicitly). The daemon PCA route
  already allowlists methods before dispatch, so the fail-loud default is
  unreachable in production (strictly safer than the old silent sticky fallback).

- test: the suite split newly cloned the `RpcFailoverClient` harness across both
  sibling files; extract `recorder`, `retryable429`, `NEVER_SIGN`, `makeClient`
  into `rpc-failover-test-helpers.ts` imported by both. Scenario-local doubles
  (readSeq, signer/contract builders, URL sets) stay local. No assertion changed.

Fast lane 600 pass / 1 skip, parity 15, tsc clean, plus an independent two-lens
equivalence review (PCA routing byte-identical; helpers byte-identical, only
import/def lines touched).


* docs(chain): drop review-plan artifacts from the split test docs (#1548 review)

Addresses otReviewAgent nit: permanent test docs should read as durable domain
language, not review history.

- rpc-failover-test-helpers.ts: remove the `PLAN §0 D1 thunks` jargon (replaced
  with the concrete "getEndpoints thunk + signPopulated callback" description) and
  the `(#1548 review)` marker.
- endpoint-stickiness.acceptance.unit.test.ts header: drop the `(#1548 item 4)`
  reference and fix the now-stale "helpers are duplicated below" note — they are
  imported from ./rpc-failover-test-helpers.js after the harness extraction.

Comment-only; no code or assertion change. The pre-existing `AC-N (round-N)`
labels in the moved block are carried over verbatim from the original suite and
left as-is (out of scope for this behavior-preserving PR).


---------
